A function for calculating the number of zooplankton per liter

This function is used to calculate the number of zooplankton per liter present in a sample that has been counted using standard sub-sampling techniques and has a known volume of water.

Usage

zoopl(count = 1, sample_vol = 2, subsample_vol = 3, trap_vol = 4)

Arguments

count

The raw count of a zooplankton species in the sub-sample

sample_vol

The volume the sample was made up to

subsample_vol

The volume of the sample that was subsampled and counted

trap_vol

The volume of water that was filtered through the trap

Value

the product of count, sample_vol/subsample_vol, and 1/trap_vol

Examples

zoopl(223, 20, 12, 30)
#> [1] 12.38889
zoopl(101, 10, 5.5, 30)
#> [1] 6.121212